Bug fixes for simpleunet

+++ MachineLearningModels/SDE_conditioned/SDE_tools.py
@@ -38,7 +38,7 @@ class GaussianDiffusion:
         scale = 1000 / self.noise_steps
         beta_start = scale * self.beta_start
         beta_end = scale * self.beta_end
-        return np.linspace(beta_start, beta_end, self.noise_steps)
+        return np.linspace(beta_start, beta_end, self.noise_steps, dtype=np.float64)
 
     def sample_timesteps(self, n):
         return torch.randint(low=1, high=self.noise_steps, size=(n,))
